+++ sibapiv3/get_aggregated_report.go
@@ -16,35 +16,35 @@ type GetAggregatedReport struct {
 	Range_ string `json:"range"`
 
 	// Number of requests for the timeframe
-	Requests int32 `json:"requests"`
+	Requests int64 `json:"requests"`
 
 	// Number of delivered emails for the timeframe
-	Delivered int32 `json:"delivered"`
+	Delivered int64 `json:"delivered"`
 
 	// Number of hardbounces for the timeframe
-	HardBounces int32 `json:"hardBounces"`
+	HardBounces int64 `json:"hardBounces"`
 
 	// Number of softbounces for the timeframe
-	SoftBounces int32 `json:"softBounces"`
+	SoftBounces int64 `json:"softBounces"`
 
 	// Number of clicks for the timeframe
-	Clicks int32 `json:"clicks"`
+	Clicks int64 `json:"clicks"`
 
 	// Number of unique clicks for the timeframe
-	UniqueClicks int32 `json:"uniqueClicks"`
+	UniqueClicks int64 `json:"uniqueClicks"`
 
 	// Number of openings for the timeframe
-	Opens int32 `json:"opens"`
+	Opens int64 `json:"opens"`
 
 	// Number of unique openings for the timeframe
-	UniqueOpens int32 `json:"uniqueOpens"`
+	UniqueOpens int64 `json:"uniqueOpens"`
 
 	// Number of complaint (spam report) for the timeframe
-	SpamReports int32 `json:"spamReports"`
+	SpamReports int64 `json:"spamReports"`
 
 	// Number of blocked contact emails for the timeframe
-	Blocked int32 `json:"blocked"`
+	Blocked int64 `json:"blocked"`
 
 	// Number of invalid emails for the timeframe
-	Invalid int32 `json:"invalid"`
+	Invalid int64 `json:"invalid"`
 }
