+++ sibapiv3/get_transac_aggregated_sms_report.go
@@ -16,29 +16,29 @@ type GetTransacAggregatedSmsReport struct {
 	Range_ string `json:"range"`
 
 	// Number of requests for the timeframe
-	Requests int32 `json:"requests"`
+	Requests int64 `json:"requests"`
 
 	// Number of delivered SMS for the timeframe
-	Delivered int32 `json:"delivered"`
+	Delivered int64 `json:"delivered"`
 
 	// Number of hardbounces for the timeframe
-	HardBounces int32 `json:"hardBounces"`
+	HardBounces int64 `json:"hardBounces"`
 
 	// Number of softbounces for the timeframe
-	SoftBounces int32 `json:"softBounces"`
+	SoftBounces int64 `json:"softBounces"`
 
 	// Number of blocked contact for the timeframe
-	Blocked int32 `json:"blocked"`
+	Blocked int64 `json:"blocked"`
 
 	// Number of unsubscription for the timeframe
-	Unsubscribed int32 `json:"unsubscribed"`
+	Unsubscribed int64 `json:"unsubscribed"`
 
 	// Number of answered SMS for the timeframe
-	Replied int32 `json:"replied"`
+	Replied int64 `json:"replied"`
 
 	// Number of accepted for the timeframe
-	Accepted int32 `json:"accepted"`
+	Accepted int64 `json:"accepted"`
 
 	// Number of rejected for the timeframe
-	Rejected int32 `json:"rejected"`
+	Rejected int64 `json:"rejected"`
 }
